pihole-mcp-server/requirements.txt
mcp>=1.0.0
google-generativeai>=0.8.0
requests>=2.31.0
pyyaml>=6.0
